requirements/production.txt
Django==1.8
argparse==1.2.1
dj-stripe==0.8.0
django-braces==1.8.1
django-filter==0.11.0
django-jsonfield==0.9.15
django-model-utils==2.3.1
django-oauth-toolkit==0.9.0
django-rest-swagger==0.3.4
djangorestframework==3.2.3
djangorestframework-filters==0.6.0
fpdf==1.7.2
gcloud==0.8.0
google-apitools==0.4.13
httplib2==0.9.2
jsonfield==1.0.3
oauth2client==1.5.2
oauthlib==1.0.1
protobuf==3.0.0-alpha-1
protorpc==0.11.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
rsa==3.3
setuptools==18.2
shortuuid==0.4.2
six==1.9.0
stripe==1.27.1
twilio==5.2.0
wsgiref==0.1.2
xlrd==0.9.4
-e git://github.com/elliothevel/django-rest-framework-stripe.git@api-updates#egg=rest-framework-stripe
